Lots and lots of changes. Version 1.0.0. See the README.

+++ library/src/main/java/com/afollestad/inquiry/callbacks/RunCallback.java
@@ -5,5 +5,5 @@ package com.afollestad.inquiry.callbacks;
  */
 public interface RunCallback {
 
-    void result(int changed);
+    void result(long changed);
 }
